move redirect_server_example.js to /test to for #12
